move textFieldEdit
