+++ libraries/text-field-edit.js
